Create Building Controller to control the level's flow

+++ Assets/Scripts/Utils/EmployeeUtils.cs
@@ -6,7 +6,7 @@ public static class EmployeeUtils
 {
     public static int DetermineRandomWeight(int minWeight, int maxWeight)
     {
-        float randomValue = Random.Range(minWeight, maxWeight);
-        return (int) randomValue;
+        int randomValue = Random.Range(minWeight, maxWeight);
+        return randomValue;
     }
 }
